Move the UI settable interface out of IJ as common deps
